backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "slurmpp-server"
version = "0.1.0"
description = "slurm++ — centralized Slurm dashboard"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "pyjwt>=2.10.0",
    "bcrypt>=4.2.0",
    "ldap3>=2.9.1",
    "apscheduler>=3.10.4",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
]

[project.scripts]
"slurm++" = "app.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "mypy>=1.13.0",
]
